# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful JSON-API for Automatic Text Recognition (ATR) developed as part of Bachelor Thesis \"FORSETE\" at NTNU Gjøvik.",
        "title": "Forsete-ATR",
        "contact": {},
        "version": "v2"
    },
    "paths": {
...
        "/forsete-atr/v2/images/{imageID}/outputs/": {
            "get": {
                "description": "Get outputs by image id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outputs"
                ],
                "summary": "Get outputs by image id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid of image",
                        "name": "imageID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "'Bearer token' must be set for valid response",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/output.Output"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/forsete-atr/v2/images/{imageID}/outputs/{outputID}/": {
            "get": {
                "description": "Get output by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outputs"
                ],
                "summary": "Get output by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid of image",
                        "name": "imageID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "uuid of output",
                        "name": "outputID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "'Bearer token' must be set for valid response",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/output.Output"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update output by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outputs"
                ],
                "summary": "Update output by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid of image",
                        "name": "imageID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "uuid of output",
                        "name": "outputID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "'Bearer token' must be set for valid response",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Body containing confirmed and data to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/output.UpdateOutputForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/output.Output"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete output by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outputs"
                ],
                "summary": "Delete output by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid of image",
                        "name": "imageID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "uuid of output",
                        "name": "outputID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "'Bearer token' must be set for valid response",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/forsete-atr/v2/images/{imageID}/outputs/{outputID}/data/": {
            "get": {
                "description": "Get output data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outputs"
                ],
                "summary": "Get output data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid of image",
                        "name": "imageID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "uuid of output",
                        "name": "outputID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "'Bearer token' must be set for valid response",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/output.ATRResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "output.ATRResponse": {
            "type": "object",
            "properties": {
                "contains": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "segment": {
                                "type": "object",
                                "properties": {
                                    "bbox": {
                                        "type": "object",
                                        "properties": {
                                            "xmax": {
                                                "type": "integer"
                                            },
                                            "xmin": {
                                                "type": "integer"
                                            },
                                            "ymax": {
                                                "type": "integer"
                                            },
                                            "ymin": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "class_label": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "object"
                                    },
                                    "orig_shape": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "polygon": {
                                        "type": "object",
                                        "properties": {
                                            "points": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "x": {
                                                            "type": "integer"
                                                        },
                                                        "y": {
                                                            "type": "integer"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    },
                                    "score": {
                                        "type": "number"
                                    }
                                }
                            },
                            "text_result": {
                                "type": "object",
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "scores": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    },
                                    "texts": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "file_name": {
                    "type": "string"
                },
                "image_name": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "output.Output": {
            "description": "Output containing id, name, format, confirmed, image_id etc.",
            "type": "object",
            "properties": {
                "confirmed": {
                    "type": "boolean"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "output.UpdateOutputForm": {
            "description": "Form containing confirmed and data associated with the update request.",
            "type": "object",
            "properties": {
                "confirmed": {
                    "type": "boolean"
                },
                "data": {
                    "$ref": "#/definitions/output.ATRResponse"
                }
            }
        },
...
        "util.ErrorResponse": {
            "description": "Json-Response on error",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}